requirements.txt
pyautogui==0.9.54
pyscreeze==0.1.30
mss==9.0.1
pillow==10.1.0
pyyaml==6.0.1
python-telegram-bot==20.6
gspread==5.12.0